Refactoring/cleanup for the observer of RIDs model design.
