+++ src/main/java/org/dsa/iot/coap/Constants.java
@@ -6,6 +6,7 @@ public class Constants {
     public static final String CONN = "conn";
     public static final String REMOTE_NAME = "RemoteDSA";
     public static final String MAIN_SERVER_NAME = "helloWorld";
+    public static final String RID_PREFIX = "__rid_";
 
     public static JsonObject makeCloseReponse(int rid) {
         JsonObject obj = new JsonObject();
